refactor: code style (#1660)
* refactor: naming

* chore: lingui extract

* refactor: migrate libs

---------
