+++ src/locales/zh-Hans/messages.po
@@ -1805,4 +1805,3 @@ msgstr "点赞"
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr "<0/> 引用了你的 <1><2/></1>"
-
