+++ src/locales/zh-Hant/messages.po
@@ -1805,4 +1805,3 @@ msgstr ""
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr ""
-
