+++ src/store/useComposeStore.ts
@@ -16,9 +16,9 @@ import { isValidRestrictionType } from '@/helpers/isValidRestrictionType.js';
 import { matchUrls } from '@/helpers/matchUrls.js';
 import { parseBlinksFromContent } from '@/helpers/parseBlinksFromContent.js';
 import { createPoll } from '@/helpers/polls.js';
-import { FrameLoader } from '@/libs/frame/Loader.js';
-import { OpenGraphLoader } from '@/libs/og/Loader.js';
 import { BlinkLoader } from '@/providers/blink/Loader.js';
+import { FrameLoader } from '@/providers/frame/Loader.js';
+import { OpenGraphLoader } from '@/providers/og/Loader.js';
 import type { Action } from '@/providers/types/Blink.js';
 import type { CompositePoll } from '@/providers/types/Poll.js';
 import type { Channel, Post } from '@/providers/types/SocialMedia.js';
